// src/DA.WI.NSGHSM.Web/ClientApp/src/app/production/input-slabs-management-root/component/hrm-input-piece-detail/view-model-extra/chemical-labels.interface.ts
export interface ChemicalLabels {
    aluminium: string;
    boron: string;
    calcium: string;
    carbon: string;
    chromium: string;
    cobalt: string;
    copper: string;
    lead: string;
    manganese: string;
    molybdenum: string;
    nickel: string;
    niobium: string;
    nitrogen: string;
    phosphorus: string;
    silicon: string;
    sulphur: string;
    tin: string;
    titanium: string;
    tungsten: string;
    vanadium: string;
}